restructure directories for clarity
