--- apps/civic_pulse/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class CivicPulseConfig(AppConfig):
-    name = 'civic_pulse'
