+++ apps/templates/agencies/agency_list.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'index.html' %}
 {% load static %}
 {% block css %}
